Refactor util.py - storage_url and retry functionality
