+++ gslib/commands/lifecycle.py
@@ -25,9 +25,9 @@ from gslib.cs_api_map import ApiSelector
 from gslib.exception import CommandException
 from gslib.exception import NO_URLS_MATCHED_TARGET
 from gslib.help_provider import CreateHelpText
+from gslib.storage_url import UrlsAreForSingleProvider
 from gslib.third_party.storage_apitools import storage_v1_messages as apitools_messages
 from gslib.translation_helper import LifecycleTranslation
-from gslib.util import UrlsAreForSingleProvider
 from gslib.utils.constants import NO_MAX
