+++ gslib/commands/rewrite.py
@@ -38,11 +38,11 @@ from gslib.storage_url import StorageUrlFromString
 from gslib.third_party.storage_apitools import storage_v1_messages as apitools_messages
 from gslib.thread_message import FileMessage
 from gslib.translation_helper import PreconditionsFromHeaders
-from gslib.util import ConvertRecursiveToFlatWildcard
 from gslib.util import GetCloudApiInstance
 from gslib.utils.constants import NO_MAX
 from gslib.utils.constants import UTF8
 from gslib.utils.system_util import StdinIterator
+from gslib.utils.text_util import ConvertRecursiveToFlatWildcard
 from gslib.utils.text_util import NormalizeStorageClass
 
 MAX_PROGRESS_INDICATOR_COLUMNS = 65
