+++ gslib/commands/stat.py
@@ -28,13 +28,13 @@ from gslib.command_argument import CommandArgument
 from gslib.cs_api_map import ApiSelector
 from gslib.exception import CommandException
 from gslib.exception import InvalidUrlError
-from gslib.ls_helper import ENCRYPTED_FIELDS
 from gslib.exception import NO_URLS_MATCHED_TARGET
-from gslib.ls_helper import UNENCRYPTED_FULL_LISTING_FIELDS
 from gslib.storage_url import ContainsWildcard
 from gslib.storage_url import StorageUrlFromString
-from gslib.util import PrintFullInfoAboutObject
 from gslib.utils.constants import NO_MAX
+from gslib.utils.ls_helper import ENCRYPTED_FIELDS
+from gslib.utils.ls_helper import PrintFullInfoAboutObject
+from gslib.utils.ls_helper import UNENCRYPTED_FULL_LISTING_FIELDS
 
 
 _SYNOPSIS = """
