+++ gslib/tests/test_acl.py
@@ -30,7 +30,7 @@ from gslib.tests.util import ObjectToURI as suri
 from gslib.tests.util import SetBotoConfigForTest
 from gslib.tests.util import unittest
 from gslib.translation_helper import AclTranslation
-from gslib.util import Retry
+from gslib.utils.retry_util import Retry
 
 PUBLIC_READ_JSON_ACL_TEXT = '"entity":"allUsers","role":"READER"'
