+++ gslib/tests/test_cors.py
@@ -24,7 +24,7 @@ import gslib.tests.testcase as testcase
 from gslib.tests.testcase.integration_testcase import SkipForS3
 from gslib.tests.util import ObjectToURI as suri
 from gslib.translation_helper import CorsTranslation
-from gslib.util import Retry
+from gslib.utils.retry_util import Retry
 
 
 @SkipForS3('CORS command is only supported for gs:// URLs')
