+++ gslib/tests/test_hashing_helper.py
@@ -23,8 +23,8 @@ import pkgutil
 from gslib.exception import CommandException
 from gslib.hashing_helper import CalculateMd5FromContents
 from gslib.hashing_helper import HashingFileUploadWrapper
+from gslib.storage_url import StorageUrlFromString
 import gslib.tests.testcase as testcase
-from gslib.util import StorageUrlFromString
 from gslib.utils.constants import TRANSFER_BUFFER_SIZE
