+++ gslib/tests/test_label.py
@@ -30,7 +30,7 @@ import gslib.tests.testcase as testcase
 from gslib.tests.testcase.integration_testcase import SkipForGS
 from gslib.tests.testcase.integration_testcase import SkipForS3
 from gslib.tests.util import ObjectToURI as suri
-from gslib.util import Retry
+from gslib.utils.retry_util import Retry
 
 KEY1 = 'key_one'
 KEY2 = 'key_two'
