+++ gslib/tests/test_mv.py
@@ -28,8 +28,8 @@ from gslib.tests.testcase.integration_testcase import SkipForS3
 from gslib.tests.util import ObjectToURI as suri
 from gslib.tests.util import SequentialAndParallelTransfer
 from gslib.tests.util import unittest
-from gslib.util import Retry
 from gslib.utils.boto_util import UsingCrcmodExtension
+from gslib.utils.retry_util import Retry
 from gslib.utils.system_util import IS_WINDOWS
