+++ gslib/tests/test_rm.py
@@ -26,7 +26,7 @@ from gslib.tests.testcase.integration_testcase import SkipForS3
 from gslib.tests.util import GenerationFromURI as urigen
 from gslib.tests.util import ObjectToURI as suri
 from gslib.tests.util import SetBotoConfigForTest
-from gslib.util import Retry
+from gslib.utils.retry_util import Retry
 
 
 class TestRm(testcase.GsUtilIntegrationTestCase):
