+++ gslib/utils/ls_helper.py
@@ -21,6 +21,11 @@ import fnmatch
 from gslib.cloud_api import EncryptionException
 from gslib.exception import CommandException
 from gslib.plurality_checkable_iterator import PluralityCheckableIterator
+from gslib.storage_url import GenerationFromUrlAndString
+from gslib.translation_helper import AclTranslation
+from gslib.utils.constants import S3_ACL_MARKER_GUID
+from gslib.utils.constants import S3_DELETE_MARKER_GUID
+from gslib.utils.constants import S3_MARKER_GUIDS
 from gslib.utils.constants import UTF8
 from gslib.utils.system_util import IS_WINDOWS
 from gslib.wildcard_iterator import StorageUrlFromString
@@ -35,20 +40,27 @@ UNENCRYPTED_FULL_LISTING_FIELDS = [
     'updated', 'timeStorageClassUpdated']
 
 
-def PrintNewLine():
-  """Default function for printing new lines between directories."""
-  print
+def MakeMetadataLine(label, value, indent=1):
+  """Returns a string with a vertically aligned label and value.
 
+  Labels of the same indentation level will start at the same column. Values
+  will all start at the same column (unless the combined left-indent and
+  label length is excessively long). If a value spans multiple lines,
+  indentation will only be applied to the first line. Example output from
+  several calls:
 
-def PrintDirHeader(bucket_listing_ref):
-  """Default function for printing headers for prefixes.
-
-  Header is printed prior to listing the contents of the prefix.
+      Label1:            Value (default indent of 1 was used)
+          Sublabel1:     Value (used indent of 2 here)
+      Label2:            Value
 
   Args:
-    bucket_listing_ref: BucketListingRef of type PREFIX.
+    label: The label to print in the first column.
+    value: The value to print in the second column.
+    indent: (4 * indent) spaces will be placed before the label.
+  Returns:
+    A string with a vertically aligned label and value.
   """
-  print '%s:' % bucket_listing_ref.url_string.encode(UTF8)
+  return '%s%s' % (((' ' * indent * 4) + label + ':').ljust(28), value)
 
 
 def PrintBucketHeader(bucket_listing_ref):  # pylint: disable=unused-argument
@@ -80,6 +92,131 @@ def PrintDirSummary(num_bytes, bucket_listing_ref):
     bucket_listing_ref: BucketListingRef of type BUCKET or PREFIX.
   """
   pass
+
+
+def PrintDirHeader(bucket_listing_ref):
+  """Default function for printing headers for prefixes.
+
+  Header is printed prior to listing the contents of the prefix.
+
+  Args:
+    bucket_listing_ref: BucketListingRef of type PREFIX.
+  """
+  print '%s:' % bucket_listing_ref.url_string.encode(UTF8)
+
+
+
+def PrintNewLine():
+  """Default function for printing new lines between directories."""
+  print
+
+
+# pylint: disable=too-many-statements
+def PrintFullInfoAboutObject(bucket_listing_ref, incl_acl=True):
+  """Print full info for given object (like what displays for gsutil ls -L).
+
+  Args:
+    bucket_listing_ref: BucketListingRef being listed.
+                        Must have ref_type OBJECT and a populated root_object
+                        with the desired fields.
+    incl_acl: True if ACL info should be output.
+
+  Returns:
+    Tuple (number of objects, object_length)
+
+  Raises:
+    Exception: if calling bug encountered.
+  """
+  url_str = bucket_listing_ref.url_string
+  storage_url = StorageUrlFromString(url_str)
+  obj = bucket_listing_ref.root_object
+
+  if (obj.metadata and S3_DELETE_MARKER_GUID in
+      obj.metadata.additionalProperties):
+    num_bytes = 0
+    num_objs = 0
+    url_str += '<DeleteMarker>'
+  else:
+    num_bytes = obj.size
+    num_objs = 1
+
+  print '%s:' % url_str.encode(UTF8)
+  if obj.timeCreated:
+    print MakeMetadataLine(
+        'Creation time', obj.timeCreated.strftime('%a, %d %b %Y %H:%M:%S GMT'))
+  if obj.updated:
+    print MakeMetadataLine(
+        'Update time', obj.updated.strftime('%a, %d %b %Y %H:%M:%S GMT'))
+  if (obj.timeStorageClassUpdated and
+      obj.timeStorageClassUpdated != obj.timeCreated):
+    print MakeMetadataLine(
+        'Storage class update time',
+        obj.timeStorageClassUpdated.strftime('%a, %d %b %Y %H:%M:%S GMT'))
+  if obj.storageClass:
+    print MakeMetadataLine('Storage class', obj.storageClass)
+  if obj.kmsKeyName:
+    print MakeMetadataLine('KMS key', obj.kmsKeyName)
+  if obj.cacheControl:
+    print MakeMetadataLine('Cache-Control', obj.cacheControl)
+  if obj.contentDisposition:
+    print MakeMetadataLine('Content-Disposition', obj.contentDisposition)
+  if obj.contentEncoding:
+    print MakeMetadataLine('Content-Encoding', obj.contentEncoding)
+  if obj.contentLanguage:
+    print MakeMetadataLine('Content-Language', obj.contentLanguage)
+  print MakeMetadataLine('Content-Length', obj.size)
+  print MakeMetadataLine('Content-Type', obj.contentType)
+  if obj.componentCount:
+    print MakeMetadataLine('Component-Count', obj.componentCount)
+  if obj.timeDeleted:
+    print MakeMetadataLine(
+        'Archived time',
+        obj.timeDeleted.strftime('%a, %d %b %Y %H:%M:%S GMT'))
+  marker_props = {}
+  if obj.metadata and obj.metadata.additionalProperties:
+    non_marker_props = []
+    for add_prop in obj.metadata.additionalProperties:
+      if add_prop.key not in S3_MARKER_GUIDS:
+        non_marker_props.append(add_prop)
+      else:
+        marker_props[add_prop.key] = add_prop.value
+    if non_marker_props:
+      print MakeMetadataLine('Metadata', '')
+      for ap in non_marker_props:
+        print MakeMetadataLine(
+            ('%s' % ap.key).encode(UTF8), ('%s' % ap.value).encode(UTF8),
+            indent=2)
+  if obj.customerEncryption:
+    if not obj.crc32c:
+      print MakeMetadataLine('Hash (crc32c)', 'encrypted')
+    if not obj.md5Hash:
+      print MakeMetadataLine('Hash (md5)', 'encrypted')
+    print MakeMetadataLine(
+        'Encryption algorithm', obj.customerEncryption.encryptionAlgorithm)
+    print MakeMetadataLine(
+        'Encryption key SHA256', obj.customerEncryption.keySha256)
+  if obj.crc32c:
+    print MakeMetadataLine('Hash (crc32c)', obj.crc32c)
+  if obj.md5Hash:
+    print MakeMetadataLine('Hash (md5)', obj.md5Hash)
+  print MakeMetadataLine('ETag', obj.etag.strip('"\''))
+  if obj.generation:
+    generation_str = GenerationFromUrlAndString(storage_url, obj.generation)
+    print MakeMetadataLine('Generation', generation_str)
+  if obj.metageneration:
+    print MakeMetadataLine('Metageneration', obj.metageneration)
+  if incl_acl:
+    # JSON API won't return acls as part of the response unless we have
+    # full control scope
+    if obj.acl:
+      print MakeMetadataLine('ACL', AclTranslation.JsonFromMessage(obj.acl))
+    elif S3_ACL_MARKER_GUID in marker_props:
+      print MakeMetadataLine('ACL', marker_props[S3_ACL_MARKER_GUID])
+    else:
+      print MakeMetadataLine('ACL', 'ACCESS DENIED')
+      print MakeMetadataLine(
+          'Note', 'You need OWNER permission on the object to read its ACL', 2)
+  return (num_objs, num_bytes)
 
 
 def PrintObject(bucket_listing_ref):
